1. 将部分无符号byte修改为int类型（解决了横向滚动时颜色突变的问题） 2. 增加一个内存查看器MemoryViewerFrame

+++ src/main/java/com/arcsinw/nesemulator/Cartridge.java
@@ -110,9 +110,9 @@ public class Cartridge {
         }
     }
 
-    public void cpuWrite(int address, byte data) {
+    public void cpuWrite(int address, int data) {
         if (address >= 0x8000 && address <= 0xFFFF) {
-            prg[address & (header.prgBanksCount > 1 ? 0x7FFF : 0x3FFF )] = data;
+            prg[address & (header.prgBanksCount > 1 ? 0x7FFF : 0x3FFF )] = (byte) data;
         }
     }
